feat(tariff): show scroll-end login popup on long tariff-updates pages
Generalize the blog scroll-to-end login trigger into a reusable
ScrollEndLoginTrigger and mount a tariff variant at the end of both the
legacy /[industryId]/tariff-updates and /chapters/[chapterSlug]/tariff-updates
pages (the longest tariff pages). The popup opens once a logged-out reader
reaches the bottom, is never enforced (click-outside closes it), shown at most
once per session, and suppressed for a few days after dismissal.

The tariff variant shares the global login-popup budget keys
(loginPopupShown / loginPopupDismissedAt) with the nav-count auto-prompt so a
reader never gets two login popups on the same tariff page.

Extract the shared bot/cooldown/storage helpers into login-popup-utils so the
auto-prompt, blog trigger, and new scroll trigger stop duplicating them. The
trigger renders only an invisible 1px sentinel plus an initially-closed popup
and skips crawlers entirely, so pages stay fully server-rendered, indexable,
and free of an interstitial for search engines.
