+++ insights-ui/src/components/login/blog-scroll-login-trigger.tsx
@@ -1,102 +1,12 @@
-'use client';
-
-import { LoginPopup } from '@/components/login/login-popup';
-import { useSession } from 'next-auth/react';
-import { useEffect, useRef, useState } from 'react';
-
-const RESHOW_COOLDOWN_MS = 3 * 24 * 60 * 60 * 1000;
-
-const SHOWN_KEY = 'blogScrollLoginShown';
-const DISMISSED_AT_KEY = 'blogScrollLoginDismissedAt';
-
-function isInCooldown(): boolean {
-  try {
-    const raw = localStorage.getItem(DISMISSED_AT_KEY);
-    if (!raw) return false;
-    const dismissedAt = Number.parseInt(raw, 10);
-    if (Number.isNaN(dismissedAt)) return false;
-    return Date.now() - dismissedAt < RESHOW_COOLDOWN_MS;
-  } catch {
-    return false;
-  }
-}
-
-function safeSetSession(key: string, value: string): void {
-  try {
-    sessionStorage.setItem(key, value);
-  } catch {
-    // sessionStorage may be unavailable (Safari private mode, etc.)
-  }
-}
-
-function safeSetLocal(key: string, value: string): void {
-  try {
-    localStorage.setItem(key, value);
-  } catch {
-    // localStorage may be unavailable
-  }
-}
+import { ScrollEndLoginTrigger } from '@/components/login/scroll-end-login-trigger';
 
 /**
- * Opens the login popup when a logged-out reader scrolls to the bottom of a blog
- * article (the "Related Articles" section).
+ * Blog-article variant of the scroll-end login trigger. Rendered just before the server-rendered
+ * "Related Articles" section so it fires when a logged-out reader reaches the bottom of an article.
  *
- * Rendered as a tiny invisible sentinel just before the server-rendered
- * `RelatedBlogs` section, so the article body and related posts stay fully
- * server-rendered — this component adds no SSR markup and does not wrap any
- * SEO-relevant content. The popup is shown at most once per browser session and
- * stays hidden for a few days after dismissal.
+ * Uses blog-scoped storage keys so the blog surface keeps its own once-per-session / cooldown
+ * budget, independent of the rest of the site.
  */
-export function BlogScrollLoginTrigger(): JSX.Element | null {
-  const { status } = useSession();
-  const sentinelRef = useRef<HTMLDivElement | null>(null);
-  const [open, setOpen] = useState<boolean>(false);
-
-  useEffect(() => {
-    if (status !== 'unauthenticated') return;
-    if (typeof IntersectionObserver === 'undefined') return;
-
-    let alreadyShown = false;
-    try {
-      alreadyShown = sessionStorage.getItem(SHOWN_KEY) === '1';
-    } catch {
-      alreadyShown = false;
-    }
-    if (alreadyShown) return;
-
-    if (isInCooldown()) return;
-
-    const sentinel = sentinelRef.current;
-    if (!sentinel) return;
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        const entry = entries[0];
-        if (entry?.isIntersecting) {
-          setOpen(true);
-          safeSetSession(SHOWN_KEY, '1');
-          observer.disconnect();
-        }
-      },
-      { rootMargin: '0px 0px -10% 0px' }
-    );
-
-    observer.observe(sentinel);
-
-    return () => {
-      observer.disconnect();
-    };
-  }, [status]);
-
-  const handleClose = (): void => {
-    setOpen(false);
-    safeSetLocal(DISMISSED_AT_KEY, String(Date.now()));
-  };
-
-  return (
-    <>
-      <div ref={sentinelRef} aria-hidden="true" className="h-px w-full" />
-      {status === 'unauthenticated' && <LoginPopup open={open} onClose={handleClose} />}
-    </>
-  );
+export function BlogScrollLoginTrigger(): JSX.Element {
+  return <ScrollEndLoginTrigger shownKey="blogScrollLoginShown" dismissedAtKey="blogScrollLoginDismissedAt" />;
 }
